work

+++ source/math/camera.c
@@ -1,11 +1,9 @@
 
 #include "camera.h"
-#include "3dmath.h"
-#include "physics.h"
 
 Camera g_cam;
 
 /* TODO leave as float for now then use fixmath int's */
-void Cam_lookpos(Vec3f *v, Camera *c)
+void Cam_lookpos(Vec3i *v, Camera *c)
 {
 }
